create dashboard

+++ frontend/src/app/component/admin/project-detail/project-detail.component.ts
@@ -78,10 +78,12 @@ export class ProjectDetailComponent implements OnInit {
   }
 
   onDelete() {
-    this._projectService.delete(this._id).subscribe((res) => {
-      console.log(res);
-      this.ngZone.run(() => this.router.navigateByUrl('/projectlist'))
-    })
+    if(confirm("Are you sure delete project ?")){
+      this._projectService.delete(this._id).subscribe((res) => {
+        console.log(res);
+        this.ngZone.run(() => this.router.navigateByUrl('/admin/project'))
+      })
+    }
   }
 
 }
